+++ internal/actions/foreach/doc.go
@@ -1,0 +1,2 @@
+// Package foreach provides functionality for executing a command on each branch in a stack.
+package foreach
