+++ internal/actions/submit/doc.go
@@ -1,0 +1,2 @@
+// Package submit provides functionality for submitting stacked branches as pull requests.
+package submit
