+++ internal/actions/submit/submit.go
@@ -211,156 +211,6 @@ func Action(ctx *app.Context, opts Options, handler Handler) error {
 	return nil
 }
 
-// prepareBranchesForSubmit prepares submission info for each branch, emitting events via handler
-func prepareBranchesForSubmit(ctx *app.Context, branches []engine.Branch, opts Options, currentBranch string, handler Handler) ([]Info, error) {
-	submissionInfos := make([]Info, 0, len(branches))
-
-	for _, branch := range branches {
-		branchName := branch.GetName()
-		status, err := branch.GetPRSubmissionStatus()
-		if err != nil {
-			return nil, err
-		}
-
-		action := status.Action
-		prNumber := status.PRNumber
-		prInfo := status.PRInfo
-
-		isCurrent := branchName == currentBranch
-
-		// Check if we should skip
-		if opts.UpdateOnly && action == "create" {
-			handler.OnEvent(BranchPlanEvent{
-				BranchName: branchName,
-				Action:     action,
-				IsCurrent:  isCurrent,
-				Skipped:    true,
-				SkipReason: "skipped, no existing PR",
-			})
-			continue
-		}
-
-		needsUpdate := status.NeedsUpdate
-		if action == "update" {
-			// Check if draft status needs to change
-			draftStatusNeedsChange := false
-			if prInfo != nil {
-				if opts.Draft && !prInfo.IsDraft() {
-					draftStatusNeedsChange = true
-				} else if opts.Publish && prInfo.IsDraft() {
-					draftStatusNeedsChange = true
-				}
-			}
-
-			needsUpdate = needsUpdate || opts.Edit || opts.Always || draftStatusNeedsChange
-
-			if !needsUpdate && !opts.Draft && !opts.Publish {
-				handler.OnEvent(BranchPlanEvent{
-					BranchName: branchName,
-					Action:     action,
-					IsCurrent:  isCurrent,
-					Skipped:    true,
-					SkipReason: status.Reason,
-				})
-				continue
-			}
-		}
-
-		// Prepare metadata
-		metadataOpts := MetadataOptions{
-			Edit:              opts.Edit && !opts.NoEdit,
-			EditTitle:         opts.EditTitle && !opts.NoEditTitle,
-			EditDescription:   opts.EditDescription && !opts.NoEditDescription,
-			NoEdit:            opts.NoEdit,
-			NoEditTitle:       opts.NoEditTitle,
-			NoEditDescription: opts.NoEditDescription,
-			Draft:             opts.Draft,
-			Publish:           opts.Publish,
-			Reviewers:         opts.Reviewers,
-			ReviewersPrompt:   opts.Reviewers == "" && opts.Edit,
-		}
-
-		metadata, err := PreparePRMetadata(branch, metadataOpts, ctx.Engine, ctx)
-		if err != nil {
-			return nil, fmt.Errorf("failed to prepare metadata for %s: %w", branchName, err)
-		}
-
-		// Get SHAs
-		headSHA, _ := branch.GetRevision()
-		parentBranchName := branch.GetParentPrecondition()
-		parentBranch := ctx.Engine.GetBranch(parentBranchName)
-		baseSHA, _ := parentBranch.GetRevision()
-
-		submissionInfo := Info{
-			BranchName: branchName,
-			Head:       branchName,
-			Base:       parentBranchName,
-			HeadSHA:    headSHA,
-			BaseSHA:    baseSHA,
-			Action:     action,
-			PRNumber:   prNumber,
-			Metadata:   metadata,
-		}
-
-		handler.OnEvent(BranchPlanEvent{
-			BranchName: branchName,
-			Action:     action,
-			IsCurrent:  isCurrent,
-			Skipped:    false,
-		})
-
-		submissionInfos = append(submissionInfos, submissionInfo)
-	}
-
-	return submissionInfos, nil
-}
-
-// getBranchesToSubmit returns the list of branches to submit based on options
-func getBranchesToSubmit(ctx *app.Context, opts Options) ([]string, error) {
-	// Get branch scope
-	branchName := opts.Branch
-	if branchName == "" {
-		currentBranch := ctx.Engine.CurrentBranch()
-		if currentBranch == nil {
-			return nil, fmt.Errorf("not on a branch and no branch specified")
-		}
-		branchName = currentBranch.GetName()
-	}
-
-	var allBranches []string
-	if opts.Stack {
-		// Include descendants and ancestors
-		branch := ctx.Engine.GetBranch(branchName)
-		stackBranches := branch.GetFullStack()
-		allBranches = make([]string, len(stackBranches))
-		for i, b := range stackBranches {
-			allBranches[i] = b.GetName()
-		}
-	} else {
-		// Just ancestors (including current branch)
-		branch := ctx.Engine.GetBranch(branchName)
-		downstackBranches := branch.GetRelativeStackDownstack()
-		allBranches = make([]string, len(downstackBranches)+1)
-		for i, b := range downstackBranches {
-			allBranches[i] = b.GetName()
-		}
-		allBranches[len(downstackBranches)] = branchName
-	}
-
-	// Remove duplicates and trunk
-	branches := []string{}
-	branchSet := make(map[string]bool)
-	for _, b := range allBranches {
-		branchObj := ctx.Engine.GetBranch(b)
-		if !branchObj.IsTrunk() && !branchSet[b] {
-			branches = append(branches, b)
-			branchSet[b] = true
-		}
-	}
-
-	return branches, nil
-}
-
 // submitBranch submits a single branch
 func submitBranch(ctx *app.Context, info Info, opts Options, handler Handler, repoOwner, repoName, remote string) error {
 	handler.OnEvent(BranchProgressEvent{
@@ -625,142 +475,4 @@ func searchSubstring(s, substr string) bool {
 		}
 	}
 	return false
-}
-
-// Event represents a feedback event from the submit action.
-// Implementations should use type switches to handle specific event types.
-type Event interface {
-	submitEvent() // marker method for type safety
-}
-
-// StackDisplayEvent indicates the initial stack visualization phase.
-// Handlers can use this to display the branches that will be processed.
-type StackDisplayEvent struct {
-	Stack    *tree.StackTree   // tree structure for rendering the stack
-	FixedMap map[string]bool   // branch -> is fixed (doesn't need restack)
-	ScopeMap map[string]string // branch -> scope
-}
-
-func (StackDisplayEvent) submitEvent() {}
-
-// RestackEvent indicates restack phase status.
-type RestackEvent struct {
-	Started   bool
-	Completed bool
-}
-
-func (RestackEvent) submitEvent() {}
-
-// PreparingEvent indicates the preparation/validation phase has started.
-type PreparingEvent struct{}
-
-func (PreparingEvent) submitEvent() {}
-
-// BranchPlanEvent indicates what will happen to each branch.
-type BranchPlanEvent struct {
-	BranchName string
-	Action     string // "create" or "update"
-	IsCurrent  bool
-	Skipped    bool
-	SkipReason string
-}
-
-func (BranchPlanEvent) submitEvent() {}
-
-// SubmissionStartEvent indicates the submission phase is beginning.
-type SubmissionStartEvent struct {
-	Branches []BranchInfo
-}
-
-func (SubmissionStartEvent) submitEvent() {}
-
-// BranchProgressEvent indicates per-branch submission progress.
-type BranchProgressEvent struct {
-	BranchName string
-	Status     BranchStatus
-	URL        string // set on success
-	Error      error  // set on failure
-}
-
-func (BranchProgressEvent) submitEvent() {}
-
-// CompletionEvent indicates the action has finished.
-type CompletionEvent struct {
-	Success bool
-	Message string // "All PRs up to date", "Dry run complete", etc.
-}
-
-func (CompletionEvent) submitEvent() {}
-
-// BranchStatus represents the status of a branch during submission.
-type BranchStatus string
-
-// BranchStatus values for tracking submission progress.
-const (
-	StatusPending    BranchStatus = "pending"
-	StatusSubmitting BranchStatus = "submitting"
-	StatusDone       BranchStatus = "done"
-	StatusError      BranchStatus = "error"
-	StatusSkipped    BranchStatus = "skipped"
-)
-
-// BranchInfo contains information about a branch for submission tracking.
-type BranchInfo struct {
-	Name     string
-	Action   string // "create" or "update"
-	PRNumber *int
-}
-
-// Handler receives events from the submit action and handles user interaction.
-// Implementations should handle events appropriately for their UI context
-// (interactive terminal, non-interactive, dashboard, etc.)
-type Handler interface {
-	// OnEvent is called for each event during submission.
-	// Handlers should use type switches to handle specific event types.
-	OnEvent(event Event)
-
-	// Confirm prompts for user confirmation when the --confirm flag is used.
-	// Returns (confirmed, error).
-	// Non-interactive handlers should return (defaultYes, nil).
-	Confirm(message string, defaultYes bool) (bool, error)
-}
-
-// ChannelHandler is a Handler that sends events to a channel.
-// Useful for async consumers like the dashboard.
-type ChannelHandler struct {
-	events chan Event
-}
-
-// NewChannelHandler creates a new ChannelHandler with a buffered channel.
-func NewChannelHandler(bufferSize int) *ChannelHandler {
-	return &ChannelHandler{
-		events: make(chan Event, bufferSize),
-	}
-}
-
-// OnEvent sends the event to the channel.
-// Non-blocking: if the channel is full, the event is dropped.
-func (h *ChannelHandler) OnEvent(e Event) {
-	select {
-	case h.events <- e:
-	default:
-		// Channel full, drop event to avoid blocking
-	}
-}
-
-// Confirm auto-confirms with the default value.
-// Dashboard operations don't support interactive confirmation.
-func (h *ChannelHandler) Confirm(_ string, defaultYes bool) (bool, error) {
-	return defaultYes, nil
-}
-
-// Events returns the event channel for reading.
-func (h *ChannelHandler) Events() <-chan Event {
-	return h.events
-}
-
-// Close closes the event channel.
-// Should be called when the action is complete.
-func (h *ChannelHandler) Close() {
-	close(h.events)
 }
