+++ internal/utils/stdin_test.go
@@ -5,13 +5,11 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-
-	"stackit.dev/stackit/internal/tui"
 )
 
 func TestReadFromStdin(t *testing.T) {
 	// Set to non-interactive mode
-	tui.SetInteractive(false)
+	SetInteractive(false)
 
 	oldStdin := os.Stdin
 	defer func() { os.Stdin = oldStdin }()
